refactor: pull components out and move add federation into federations table
