+++ apps/gateway-ui/src/App.tsx
@@ -1,32 +1,14 @@
 import React, { useState, useEffect, useMemo } from 'react';
-import {
-  Box,
-  Button,
-  Heading,
-  Text,
-  Flex,
-  useTheme,
-  CircularProgress,
-  CircularProgressLabel,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalCloseButton,
-  ModalBody,
-  Table,
-  Thead,
-  Tbody,
-  Tr,
-  Th,
-  Td,
-} from '@chakra-ui/react';
+import { Box, Heading, Flex, useTheme } from '@chakra-ui/react';
 import { GatewayInfo, FederationInfo } from '@fedimint/types';
 import { ConnectFederation } from './components';
 import { GatewayApi } from './GatewayApi';
 import { ApiProvider } from './ApiProvider';
 import { Wrapper, Login } from '@fedimint/ui';
 import { useTranslation } from '@fedimint/utils';
-import { FederationsTable } from './components/fedsTable/FederationsTable';
+import { FederationsTable } from './components/federations/FederationsTable';
+import { Loading } from './components/Loading';
+import { Error } from './components/Error';
 
 export const App = React.memo(function Admin(): JSX.Element {
   const gateway = useMemo(() => new GatewayApi(), []);
@@ -99,52 +81,11 @@ export const App = React.memo(function Admin(): JSX.Element {
 
   const content = useMemo(() => {
     if (loading) {
-      return (
-        <Flex
-          bgColor={theme.colors.white}
-          justifyContent='center'
-          alignItems='center'
-        >
-          <CircularProgress
-            isIndeterminate={true}
-            color={theme.colors.blue[600]}
-            size='240px'
-            thickness='8px'
-            capIsRound={true}
-          >
-            <CircularProgressLabel
-              fontSize='md'
-              fontWeight='500'
-              color={theme.colors.gray[600]}
-              fontFamily={theme.fonts.body}
-              textAlign='center'
-              width='150px'
-            >
-              {t('admin.fetch-info-modal-text')}
-            </CircularProgressLabel>
-          </CircularProgress>
-        </Flex>
-      );
+      return <Loading />;
     }
-
     if (error) {
-      return (
-        <Flex
-          direction='column'
-          align='center'
-          width='100%'
-          paddingTop='10vh'
-          paddingX='4'
-          textAlign='center'
-        >
-          <Heading size='lg' marginBottom='4'>
-            {t('common.error')}
-          </Heading>
-          <Text fontSize='md'>{error}</Text>
-        </Flex>
-      );
+      return <Error error={error} />;
     }
-
     if (!authenticated) {
       return (
         <Login
@@ -175,32 +116,23 @@ export const App = React.memo(function Admin(): JSX.Element {
           >
             {t('header.title')}
           </Heading>
-          <Button onClick={() => setShowConnectFed(!showConnectFed)}>
-            {t('connect-federation.connect-federation-button')}
-          </Button>
         </Flex>
-        <Modal isOpen={showConnectFed} onClose={() => setShowConnectFed(false)}>
-          <ModalOverlay />
-          <ModalContent>
-            <ModalCloseButton />
-            <br />
-            <ModalBody>
-              <ConnectFederation
-                renderConnectedFedCallback={(federation: FederationInfo) => {
-                  setGatewayInfo({
-                    ...gatewayInfo,
-                    federations: [...gatewayInfo.federations, federation],
-                  });
-                  setShowConnectFed(false);
-                }}
-              />
-            </ModalBody>
-          </ModalContent>
-        </Modal>
+        <ConnectFederation
+          isOpen={showConnectFed}
+          onClose={() => setShowConnectFed(false)}
+          renderConnectedFedCallback={(federation: FederationInfo) => {
+            setGatewayInfo({
+              ...gatewayInfo,
+              federations: [...gatewayInfo.federations, federation],
+            });
+            setShowConnectFed(false);
+          }}
+        />
         <FederationsTable
           federations={gatewayInfo.federations}
           onDeposit={handleDeposit}
           onWithdraw={handleWithdraw}
+          onConnectFederation={() => setShowConnectFed(true)}
         />
       </Box>
     );
